ComplaintController 
민원선택과 게시판 경로 잡아줌
(게시글 삭제 구현)

jsp 및 서비스&서비스Impl
스프링 AOP 방식에 맞게 적용

ComplaintVO
변경된 DB에 따른 멤버이름 변경

유저메뉴와 푸터(userMenu, footer)
jsp에 추가(민원선택과 민원게시판jsp)

+++ Aptogether/src/main/java/org/aptogether/domain/ComplaintVO.java
@@ -9,6 +9,6 @@ public class ComplaintVO {
 	String comptel;
 	String compcontent;
 	String compdate;
-	int compbno;	
+	long compbno;	
 	
 }
